+++ tests/modules/write.test.ts
@@ -33,8 +33,8 @@ import {
   DanModuleInfo,
   ECDSA_OWNERSHIP_MODULE_ADDRESSES_BY_VERSION,
   NodeWallet,
-  type PolicyWithoutSessionKey,
-  createDistributedSession,
+  type PolicyLeaf,
+  createDelegatedSession,
   createECDSAOwnershipValidationModule,
   createMultiChainValidationModule,
   createSessionKeyManagerModule,
@@ -176,14 +176,11 @@ describe("Modules:Write", () => {
 
   // User must be connected with a wallet to grant permissions
   test("should create a single session on behalf of a user", async () => {
-    const { sessionKeyAddress, sessionStorageClient } =
-      await createSessionKeyEOA(smartAccountThree, chain)
 
     const { wait, session } = await createSession(
       smartAccountThree,
       [
         {
-          sessionKeyAddress,
           contractAddress: nftAddress,
           functionSelector: "safeMint(address)",
           rules: [
@@ -200,7 +197,7 @@ describe("Modules:Write", () => {
           valueLimit: 0n
         }
       ],
-      sessionStorageClient,
+      undefined,
       withSponsorship
     )
 
@@ -1463,7 +1460,7 @@ describe("Modules:Write", () => {
 
   test("should create and use an DAN session on behalf of a user (abstracted)", async () => {
 
-    const policy: PolicyWithoutSessionKey[] = [
+    const policy: PolicyLeaf[] = [
       {
         contractAddress: nftAddress,
         functionSelector: "safeMint(address)",
@@ -1479,7 +1476,7 @@ describe("Modules:Write", () => {
       }
     ]
 
-    const { wait } = await createDistributedSession({ smartAccountClient: smartAccount, policy })
+    const { wait } = await createDelegatedSession({ smartAccountClient: smartAccount, policy })
 
     const { success } = await wait()
     expect(success).toBe("true")
@@ -1503,7 +1500,7 @@ describe("Modules:Write", () => {
         chainId
       },
       smartAccountAddress,
-      "DAN"
+      "DELEGATED"
     )
 
     const { wait: waitForMint } = await smartAccountWithSession.sendTransaction(
@@ -1546,7 +1543,7 @@ describe("Modules:Write", () => {
     const duration = 60 * 60
 
     // Get the session key from the dan network
-    const danSessionData = await getDANSessionKey({
+    const danModuleInfo = await getDANSessionKey({
       smartAccountClient: smartAccount,
       browserWallet: new NodeWallet(walletClient),
       duration
@@ -1556,7 +1553,7 @@ describe("Modules:Write", () => {
     const policy: Policy[] = [{
       contractAddress: nftAddress,
       functionSelector: "safeMint(address)",
-      sessionKeyAddress: danSessionData.sessionKeyEOA, // Add the session key address from DAN
+      sessionKeyAddress: danModuleInfo.sessionKeyEOA, // Add the session key address from DAN
       rules: [
         {
           offset: RuleHelpers.OffsetByIndex(0),
@@ -1573,7 +1570,7 @@ describe("Modules:Write", () => {
 
     // Create the session data using the information retrieved from DAN. Keep the danModuleInfo for later use in a session leaf
     const { data: policyData, sessionIDInfo: sessionIDs } =
-      await sessionsModule.createSessionData(policy.map(p => createABISessionDatum({ ...p, danModuleInfo: { ...danSessionData, chainId } })))
+      await sessionsModule.createSessionData(policy.map(p => createABISessionDatum({ ...p, danModuleInfo })))
 
     // Cconstruct the session transaction
     const permitTx = {
@@ -1646,10 +1643,13 @@ describe("Modules:Write", () => {
     // Assume we know that the relevant session leaf to the transaction is the last one...
     const allLeaves = await memoryStore.getAllSessionData();
     const relevantLeaf = allLeaves[allLeaves.length - 1];
+
     const sessionID = relevantLeaf.sessionID;
+    // OR 
+    const sameSessionID = sessionIDs[0]; // Usually only available when the session is created
 
     const nftBalanceBefore = await checkBalance(smartAccountAddress, nftAddress);
-    // Now use the sessionID and the DAN module info to send the transaction
+    // Now use the sessionID to send the transaction
     const { wait: waitForMint } = await unconnectedSmartAccount.sendTransaction(nftMintTx, {
       ...withSponsorship,
       params: { sessionID }